Refactored CharacteristicInput to use child component for each radio input
